+++ src/graphql.rs
@@ -1,9 +1,14 @@
 mod log;
 mod network;
 
-use crate::storage::{
-    lower_closed_bound_key, upper_closed_bound_key, upper_open_bound_key, Database, Direction,
-    KeyValue, RawEventStore,
+use std::net::IpAddr;
+
+use crate::{
+    ingestion::EventFilter,
+    storage::{
+        lower_closed_bound_key, upper_closed_bound_key, upper_open_bound_key, Database, Direction,
+        KeyValue, RawEventStore,
+    },
 };
 use anyhow::anyhow;
 use async_graphql::{
@@ -18,31 +23,76 @@ const TIMESTAMP_SIZE: usize = 8;
 pub struct Query(log::LogQuery, network::NetworkQuery);
 
 #[derive(InputObject)]
-struct RawEventFilterInput {
+pub struct RawEventFilterInput {
     time: Option<TimeRange>,
     source: String,
-    orig_addr: Option<IpRange>,
-    resp_addr: Option<IpRange>,
-    orig_port: Option<PortRange>,
-    resp_port: Option<PortRange>,
+    kind: Option<String>,
+    pub orig_addr: Option<IpRange>,
+    pub resp_addr: Option<IpRange>,
+    pub orig_port: Option<PortRange>,
+    pub resp_port: Option<PortRange>,
 }
 
-#[derive(InputObject, Debug, Default)]
+impl RawEventFilterInput {
+    fn check(
+        &self,
+        o_a: Option<IpAddr>,
+        r_a: Option<IpAddr>,
+        o_p: Option<u16>,
+        r_p: Option<u16>,
+    ) -> Result<bool> {
+        if let Some(orig_addr) = &self.orig_addr {
+            if let Some(o_a) = o_a {
+                if (o_a >= orig_addr.end.parse::<IpAddr>()?)
+                    || (o_a < orig_addr.start.parse::<IpAddr>()?)
+                {
+                    return Ok(false);
+                };
+            }
+        }
+        if let Some(resp_addr) = &self.resp_addr {
+            if let Some(r_a) = r_a {
+                if (r_a >= resp_addr.end.parse::<IpAddr>()?)
+                    || (r_a < resp_addr.start.parse::<IpAddr>()?)
+                {
+                    return Ok(false);
+                };
+            }
+        }
+        if let Some(orig_port) = &self.orig_port {
+            if let Some(o_p) = o_p {
+                if (o_p >= orig_port.end) || (o_p < orig_port.start) {
+                    return Ok(false);
+                };
+            }
+        }
+        if let Some(resp_port) = &self.resp_port {
+            if let Some(r_p) = r_p {
+                if (r_p >= resp_port.end) || (r_p < resp_port.start) {
+                    return Ok(false);
+                };
+            }
+        }
+        Ok(true)
+    }
+}
+
+#[derive(InputObject)]
 struct TimeRange {
-    start: DateTime<Utc>,
-    end: DateTime<Utc>,
+    start: Option<DateTime<Utc>>,
+    end: Option<DateTime<Utc>>,
 }
 
-#[derive(InputObject, Debug)]
-struct IpRange {
-    start: String,
-    end: String,
+#[derive(InputObject)]
+pub struct IpRange {
+    pub start: String,
+    pub end: String,
 }
 
-#[derive(InputObject, Debug)]
-struct PortRange {
-    start: u16,
-    end: u16,
+#[derive(InputObject)]
+pub struct PortRange {
+    pub start: u16,
+    pub end: u16,
 }
 
 pub trait FromKeyValue<T>: Sized {
@@ -62,13 +112,12 @@ pub fn schema(database: Database) -> Schema {
 /// Maximum size: 100.
 const MAXIMUM_PAGE_SIZE: usize = 100;
 
-#[allow(clippy::too_many_arguments)]
-fn load_connection_log<'c, N, I, T>(
+#[allow(clippy::too_many_lines, clippy::too_many_arguments)]
+fn load_connection<'c, N, I, T>(
     store: &RawEventStore<'c>,
     key_prefix: &[u8],
     iter_builder: fn(&RawEventStore<'c>, &[u8], &[u8], Direction) -> I,
-    start: Option<DateTime<Utc>>,
-    end: Option<DateTime<Utc>>,
+    filter: &RawEventFilterInput,
     after: Option<String>,
     before: Option<String>,
     first: Option<usize>,
@@ -77,6 +126,7 @@ fn load_connection_log<'c, N, I, T>(
 where
     N: FromKeyValue<T> + OutputType,
     I: Iterator<Item = anyhow::Result<(Box<[u8]>, T)>> + 'c,
+    T: EventFilter,
 {
     let (records, has_previous, has_next) = if let Some(before) = before {
         if after.is_some() {
@@ -85,6 +135,12 @@ where
         if first.is_some() {
             return Err("'before' and 'first' cannot be specified simultaneously".into());
         }
+        let (start, end) = if let Some(ref time) = filter.time {
+            (time.start, time.end)
+        } else {
+            (None, None)
+        };
+
         let last = last.unwrap_or(MAXIMUM_PAGE_SIZE).min(MAXIMUM_PAGE_SIZE);
         let cursor = base64::decode(before)?;
         let time = upper_closed_bound_key(key_prefix, end);
@@ -103,7 +159,7 @@ where
                 iter.next();
             }
         }
-        let (mut records, has_previous) = collect_records(iter, last)?;
+        let (mut records, has_previous) = collect_records(iter, last, filter)?;
         records.reverse();
         (records, has_previous, false)
     } else if let Some(after) = after {
@@ -113,6 +169,12 @@ where
         if last.is_some() {
             return Err("'after' and 'last' cannot be specified simultaneously".into());
         }
+        let (start, end) = if let Some(ref time) = filter.time {
+            (time.start, time.end)
+        } else {
+            (None, None)
+        };
+
         let first = first.unwrap_or(MAXIMUM_PAGE_SIZE).min(MAXIMUM_PAGE_SIZE);
         let cursor = base64::decode(after)?;
         let time = lower_closed_bound_key(key_prefix, start);
@@ -131,12 +193,18 @@ where
                 iter.next();
             }
         }
-        let (records, has_next) = collect_records(iter, first)?;
+        let (records, has_next) = collect_records(iter, first, filter)?;
         (records, false, has_next)
     } else if let Some(last) = last {
         if first.is_some() {
             return Err("first and last cannot be used together".into());
         }
+        let (start, end) = if let Some(ref time) = filter.time {
+            (time.start, time.end)
+        } else {
+            (None, None)
+        };
+
         let last = last.min(MAXIMUM_PAGE_SIZE);
         let iter = iter_builder(
             store,
@@ -144,10 +212,16 @@ where
             &lower_closed_bound_key(key_prefix, start),
             Direction::Reverse,
         );
-        let (mut records, has_previous) = collect_records(iter, last)?;
+        let (mut records, has_previous) = collect_records(iter, last, filter)?;
         records.reverse();
         (records, has_previous, false)
     } else {
+        let (start, end) = if let Some(ref time) = filter.time {
+            (time.start, time.end)
+        } else {
+            (None, None)
+        };
+
         let first = first.unwrap_or(MAXIMUM_PAGE_SIZE).min(MAXIMUM_PAGE_SIZE);
         let iter = iter_builder(
             store,
@@ -155,7 +229,7 @@ where
             &upper_open_bound_key(key_prefix, end),
             Direction::Forward,
         );
-        let (records, has_next) = collect_records(iter, first)?;
+        let (records, has_next) = collect_records(iter, first, filter)?;
         (records, false, has_next)
     };
 
@@ -172,118 +246,28 @@ where
     Ok(connection)
 }
 
-#[allow(clippy::too_many_arguments, clippy::too_many_lines)]
-fn load_connection<'c, N, I, T>(
-    store: &RawEventStore<'c>,
-    key_prefix: &[u8],
-    iter_builder: fn(&RawEventStore<'c>, &[u8], &[u8], Direction) -> I,
-    filter: RawEventFilterInput,
-    after: Option<String>,
-    before: Option<String>,
-    first: Option<usize>,
-    last: Option<usize>,
-) -> Result<Connection<String, N>>
-where
-    N: FromKeyValue<T> + OutputType,
-    I: Iterator<Item = anyhow::Result<(Box<[u8]>, T)>> + 'c,
-    T: std::clone::Clone,
-{
-    let (records, has_previous, has_next) = if let Some(before) = before {
-        if after.is_some() {
-            return Err("cannot use both `after` and `before`".into());
-        }
-        if first.is_some() {
-            return Err("'before' and 'first' cannot be specified simultaneously".into());
-        }
-        let last = last.unwrap_or(MAXIMUM_PAGE_SIZE).min(MAXIMUM_PAGE_SIZE);
-        let cursor = base64::decode(before)?;
-
-        let mut iter = iter_builder(
-            store,
-            &cursor,
-            &upper_open_bound_key(key_prefix, Some(filter.time.unwrap_or_default().start)),
-            Direction::Reverse,
-        )
-        .peekable();
-        if let Some(Ok((key, _))) = iter.peek() {
-            if key.as_ref() == cursor {
-                iter.next();
-            }
-        }
-        let (mut records, has_previous) = collect_records(iter, last)?;
-        records.reverse();
-        (records, has_previous, false)
-    } else if let Some(after) = after {
-        if before.is_some() {
-            return Err("cannot use both `after` and `before`".into());
-        }
-        if last.is_some() {
-            return Err("'after' and 'last' cannot be specified simultaneously".into());
-        }
-        let _first = first.unwrap_or(MAXIMUM_PAGE_SIZE).min(MAXIMUM_PAGE_SIZE);
-        let cursor = base64::decode(after)?;
-        let mut iter = iter_builder(
-            store,
-            &cursor,
-            &upper_open_bound_key(key_prefix, Some(filter.time.unwrap_or_default().end)),
-            Direction::Forward,
-        )
-        .peekable();
-        if let Some(Ok((key, _))) = iter.peek() {
-            if key.as_ref() == cursor {
-                iter.next();
-            }
-        }
-        let (records, has_next) = collect_records(iter, last.unwrap())?;
-        (records, false, has_next)
-    } else if let Some(last) = last {
-        if first.is_some() {
-            return Err("first and last cannot be used together".into());
-        }
-        let last = last.min(MAXIMUM_PAGE_SIZE);
-        let iter = iter_builder(
-            store,
-            &upper_closed_bound_key(key_prefix, None),
-            &upper_open_bound_key(key_prefix, Some(filter.time.unwrap().start)),
-            Direction::Reverse,
-        );
-        let (mut records, has_previous) = collect_records(iter, last)?;
-        records.reverse();
-        (records, has_previous, false)
-    } else {
-        let first = first.unwrap_or(MAXIMUM_PAGE_SIZE).min(MAXIMUM_PAGE_SIZE);
-        let iter = iter_builder(
-            store,
-            key_prefix,
-            &upper_open_bound_key(key_prefix, Some(filter.time.unwrap().end)),
-            Direction::Forward,
-        );
-        let (records, has_next) = collect_records(iter, first)?;
-        (records, false, has_next)
-    };
-
-    let mut connection: Connection<String, N> = Connection::new(has_previous, has_next);
-    connection.edges = records
-        .into_iter()
-        .map(|(key, node)| {
-            Edge::new(
-                base64::encode(&key),
-                N::from_key_value(&key, node).expect("failed to convert value"),
-            )
-        })
-        .collect();
-    Ok(connection)
-}
-
-fn collect_records<I, T>(mut iter: I, size: usize) -> Result<(Vec<KeyValue<T>>, bool)>
+fn collect_records<I, T>(
+    mut iter: I,
+    size: usize,
+    filter: &RawEventFilterInput,
+) -> Result<(Vec<KeyValue<T>>, bool)>
 where
     I: Iterator<Item = anyhow::Result<(Box<[u8]>, T)>>,
+    T: EventFilter,
 {
     let mut records = Vec::with_capacity(size);
     let mut has_more = false;
     while let Some(item) = iter.next() {
         let item = item.map_err(|e| format!("failed to read database: {}", e))?;
-        records.push(item);
+        match filter.check(
+            item.1.orig_addr(),
+            item.1.resp_addr(),
+            item.1.orig_port(),
+            item.1.resp_port(),
+        ) {
+            Ok(true) => records.push(item),
+            Ok(false) | Err(_) => {}
+        }
         if records.len() == size {
             has_more = iter.next().is_some();
             break;